requirements.txt
rdkit~=2022.3.5
rdkit
pandas~=1.4.4
colorama~=0.4.5
pyarrow
dgl~=0.9.1
numpy~=1.23.3
networkx~=2.8.6
scipy~=1.9.1
tqdm~=4.64.1
attrs~=21.4.0
pyyaml~=6.0
pydantic~=1.9.0
pytest~=7.1.2
jupyter
hydra
torch~=1.12.1
yaml~=0.2.5
omegaconf~=2.2.3